refactor: use CDNs instead of copy/pasting JS libs
In order to profit from library updates (including security updates), we need to find a way to make the JS libs easier exchangeable. By loading them from CDNs, we make a first step towards that (with some more configuration we could profit from RenovateBot). The all-over-the-place inclusion of JS libraries is still problematic and we need further changes to untangle this.