chore: only use Tailwind moving forward
This change removes all old custom CSS (in SCSS files etc.). Since almost all parts of the website will be redesigned, it's okay to remove the old CSS already and start with a blank slate using Tailwind. With Tailwind we are able to move fast and this way we can be sure that no legacy CSS slips into our styling.

This change also updates the build and dev scripts to use the TailwindCLI.