feat: add basic highlighted datasets
We want to incorporate a "Wimmelbild" for users to explore open and closed datasets in Berlin. There will be lots of visual work to do, but we'll start with the accessible basics. Presumably, the datasets we want to shed light on can be thought of a a list. Therefore we use the accessible "list" role. Every visual changes following should always work with this (or a similar) markup.

Note that we haven't included the actual datasets yet, for now we are using 3 example entries that represent the possible attributes.