chore: only have one h1 in news articles
This is done by making a h2 out of every h1 that was at the beginning of a news article. We do this because there can only be one h1 per page. That is already achieved through the title front matter that is transformed to a h1 in the news layout.

Note that this doesn't touch subsequent heading levels. So no h2 becomes a h3 with this change. This is because in most of the articles this would be semantically incorrect anyways.